[NOM-18] - Refactored file/class names
