+++ lib/main.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:nomad/constants.dart';
-import 'package:nomad/widgets/dynamic_search_card.dart';
+import 'package:nomad/widgets/country_searchbar.dart';
 
 void main() {
   runApp(const MyApp());
