+++ lib/widgets/country_searchbar.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 
 import '../domain/country.dart';
+import 'country_card.dart';
 
 class CountrySearchBar extends StatefulWidget {
   CountrySearchBar({super.key, required this.allCountries});
@@ -86,59 +87,11 @@ class _CountrySearchBarState extends State<CountrySearchBar> {
           onTap: () {
             searchController.openView();
           },
-          onChanged: (_) {
-            searchController.openView();
-          },
         );
       },
       suggestionsBuilder: (BuildContext context, SearchController searchController) {
-        return [Text('Not used - this allows us to provide our own ListView.Builder above for lazy loading...')];
+        return [Text('required but not used - this allows us to provide our own ListView.Builder above for lazy loading...')];
       }
     );
   }
 }
-
-class CountryCard extends StatelessWidget {
-  const CountryCard({super.key, required this.country, required this.onTap});
-
-  final Country country;
-  final VoidCallback onTap;
-
-  @override
-  Widget build(BuildContext context) {
-    return Column(
-      children: [
-        GestureDetector(
-          onTap: onTap,
-          child: Card(
-            margin: EdgeInsets.all(0.0),
-            shape: ContinuousRectangleBorder(),
-            child: Padding(
-              padding: const EdgeInsets.all(8.0),
-              child: Row(
-                children: [
-                  Icon(country.getIcon),
-                  SizedBox(width: 12,),
-                  Expanded(
-                    child: Column(
-                      crossAxisAlignment: CrossAxisAlignment.start,
-                      children: [
-                        Text(country.getName, style: TextStyle(fontWeight: FontWeight.bold, fontSize: 18),),
-                        Text(country.getDescription, overflow: TextOverflow.ellipsis,)
-                      ],
-                    ),
-                  ),
-                ],
-              ),
-            ),
-          ),
-        ),
-        Divider(
-          color: Colors.black,
-          height: 0,
-        )
-      ],
-    );
-  }
-}
-
